New version 3.0
Included in this version

Removal of the resource_name_convention
Correct the diff generation to avoid recreation when not needed
Added import support
