+++ .vscode/launch.json
@@ -1,25 +1,35 @@
 {
 	"version": "0.2.0",
 	"configurations": [
-	
-	{
-		"name": "Launch test package",
-		"type": "go",
-		"request": "launch",
-		"mode": "test",
-		//"program": "${workspaceFolder}/main.go"
-		"args": [
-			"check-go-version",
-			"false"
-		],
-		"program": "${file}"
-	},
- 		{
+		{
+			"name": "Debug provider",
+			"type": "go",
+			"request": "launch",
+			"mode": "debug",
+			"program": "${workspaceFolder}",
+			"args": [
+				"--",
+				"debug"
+			]
+		},
+		{
+			"name": "Launch test package",
+			"type": "go",
+			"request": "launch",
+			"mode": "test",
+			//"program": "${workspaceFolder}/main.go"
+			"args": [
+				"check-go-version",
+				"false"
+			],
+			"program": "${file}"
+		},
+		{
 			"name": "Attach",
 			"type": "go",
 			"request": "attach",
-            "mode": "local",
-            "processId": 23310   // pgrep terraform, put the latest number. Add time.Sleep(50 * time.Second) before the break point
-        }
+			"mode": "local",
+			"processId": 23310 // pgrep terraform, put the latest number. Add time.Sleep(50 * time.Second) before the break point
+		}
 	]
 }
