+++ azurecaf/resource_name.go
@@ -2,127 +2,42 @@ package azurecaf
 
 import (
 	"context"
+	b64 "encoding/base64"
 	"fmt"
-	"regexp"
 	"strings"
 
+	"github.com/google/martian/v3/log"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 )
-
-func resourceNameV2() *schema.Resource {
-	resourceMapsKeys := make([]string, 0, len(ResourceDefinitions))
-	for k := range ResourceDefinitions {
-		resourceMapsKeys = append(resourceMapsKeys, k)
-	}
-	return &schema.Resource{
-		Schema: map[string]*schema.Schema{
-			"name": {
-				Type:     schema.TypeString,
-				Optional: true,
-				ForceNew: true,
-				Default:  "",
-			},
-			"prefixes": {
-				Type: schema.TypeList,
-				Elem: &schema.Schema{
-					Type:         schema.TypeString,
-					ValidateFunc: validation.NoZeroValues,
-				},
-				Optional: true,
-				ForceNew: true,
-			},
-			"suffixes": {
-				Type: schema.TypeList,
-				Elem: &schema.Schema{
-					Type:         schema.TypeString,
-					ValidateFunc: validation.NoZeroValues,
-				},
-				Optional: true,
-				ForceNew: true,
-			},
-			"random_length": {
-				Type:         schema.TypeInt,
-				Optional:     true,
-				ForceNew:     true,
-				ValidateFunc: validation.IntAtLeast(0),
-				Default:      0,
-			},
-			"result": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"results": {
-				Type: schema.TypeMap,
-				Elem: &schema.Schema{
-					Type: schema.TypeString,
-				},
-				Computed: true,
-			},
-			"separator": {
-				Type:     schema.TypeString,
-				Optional: true,
-				ForceNew: true,
-				Default:  "-",
-			},
-			"clean_input": {
-				Type:     schema.TypeBool,
-				Optional: true,
-				ForceNew: true,
-				Default:  true,
-			},
-			"passthrough": {
-				Type:     schema.TypeBool,
-				Optional: true,
-				ForceNew: true,
-				Default:  false,
-			},
-			"resource_type": {
-				Type:         schema.TypeString,
-				Optional:     true,
-				ValidateFunc: validation.StringInSlice(resourceMapsKeys, false),
-				ForceNew:     true,
-			},
-			"resource_types": {
-				Type: schema.TypeList,
-				Elem: &schema.Schema{
-					Type:         schema.TypeString,
-					ValidateFunc: validation.StringInSlice(resourceMapsKeys, false),
-				},
-				Optional: true,
-				ForceNew: true,
-			},
-			"random_seed": {
-				Type:     schema.TypeInt,
-				Optional: true,
-				ForceNew: true,
-			},
-		},
-	}
-}
-
-func resourceNameStateUpgradeV2(ctx context.Context, rawState map[string]interface{}, meta interface{}) (map[string]interface{}, error) {
-	rawState["use_slug"] = true
-
-	return rawState, nil
-}
 
 func resourceName() *schema.Resource {
 	resourceMapsKeys := make([]string, 0, len(ResourceDefinitions))
 	for k := range ResourceDefinitions {
 		resourceMapsKeys = append(resourceMapsKeys, k)
 	}
+	log.Debugf("Initialized provider")
 
 	return &schema.Resource{
-		Create:        resourceNameCreate,
-		Read:          schema.Noop,
-		Delete:        schema.RemoveFromState,
-		SchemaVersion: 3,
+		Create: resourceNameCreate,
+		Update: resourceNameUpdate,
+		Read:   resourceNameRead,
+		Delete: schema.RemoveFromState,
+		Importer: &schema.ResourceImporter{
+
+			State: importState,
+		},
+		SchemaVersion: 4,
 		StateUpgraders: []schema.StateUpgrader{
 			{
 				Type:    resourceNameV2().CoreConfigSchema().ImpliedType(),
 				Upgrade: resourceNameStateUpgradeV2,
 				Version: 2,
+			},
+			{
+				Type:    resourceNameV3().CoreConfigSchema().ImpliedType(),
+				Upgrade: resourceNameStateUpgradeV3,
+				Version: 3,
 			},
 		},
 
@@ -130,7 +45,7 @@ func resourceName() *schema.Resource {
 			"name": {
 				Type:     schema.TypeString,
 				Optional: true,
-				ForceNew: true,
+				ForceNew: false,
 				Default:  "",
 			},
 			"prefixes": {
@@ -140,7 +55,7 @@ func resourceName() *schema.Resource {
 					ValidateFunc: validation.NoZeroValues,
 				},
 				Optional: true,
-				ForceNew: true,
+				ForceNew: false,
 			},
 			"suffixes": {
 				Type: schema.TypeList,
@@ -149,12 +64,12 @@ func resourceName() *schema.Resource {
 					ValidateFunc: validation.NoZeroValues,
 				},
 				Optional: true,
-				ForceNew: true,
+				ForceNew: false,
 			},
 			"random_length": {
 				Type:         schema.TypeInt,
 				Optional:     true,
-				ForceNew:     true,
+				ForceNew:     false,
 				ValidateFunc: validation.IntAtLeast(0),
 				Default:      0,
 			},
@@ -172,26 +87,26 @@ func resourceName() *schema.Resource {
 			"separator": {
 				Type:     schema.TypeString,
 				Optional: true,
-				ForceNew: true,
+				ForceNew: false,
 				Default:  "-",
 			},
 			"clean_input": {
 				Type:     schema.TypeBool,
 				Optional: true,
-				ForceNew: true,
+				ForceNew: false,
 				Default:  true,
 			},
 			"passthrough": {
 				Type:     schema.TypeBool,
 				Optional: true,
-				ForceNew: true,
+				ForceNew: false,
 				Default:  false,
 			},
 			"resource_type": {
 				Type:         schema.TypeString,
 				Optional:     true,
 				ValidateFunc: validation.StringInSlice(resourceMapsKeys, false),
-				ForceNew:     true,
+				ForceNew:     false,
 			},
 			"resource_types": {
 				Type: schema.TypeList,
@@ -214,232 +129,17 @@ func resourceName() *schema.Resource {
 				Default:  true,
 			},
 		},
+		CustomizeDiff: getDifference,
 	}
 }
 
-func resourceNameCreate(d *schema.ResourceData, meta interface{}) error {
-	return resourceNameRead(d, meta)
+func importState(d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
+	d.Set("result", d.Id())
+	d.SetId("imported")
+	return []*schema.ResourceData{d}, nil
 }
 
-func resourceNameRead(d *schema.ResourceData, meta interface{}) error {
-	return getNameResult(d, meta)
-}
-
-func resourceNameDelete(d *schema.ResourceData, meta interface{}) error {
-	return nil
-}
-
-func cleanSlice(names []string, resourceDefinition *ResourceStructure) []string {
-	for i, name := range names {
-		names[i] = cleanString(name, resourceDefinition)
-	}
-	return names
-}
-
-func cleanString(name string, resourceDefinition *ResourceStructure) string {
-	myRegex, _ := regexp.Compile(resourceDefinition.RegEx)
-	return myRegex.ReplaceAllString(name, "")
-}
-
-func concatenateParameters(separator string, parameters ...[]string) string {
-	elems := []string{}
-	for _, items := range parameters {
-		for _, item := range items {
-			if len(item) > 0 {
-				elems = append(elems, []string{item}...)
-			}
-		}
-	}
-	return strings.Join(elems, separator)
-}
-
-func getResource(resourceType string) (*ResourceStructure, error) {
-	if resourceKey, existing := ResourceMaps[resourceType]; existing {
-		resourceType = resourceKey
-	}
-	if resource, resourceFound := ResourceDefinitions[resourceType]; resourceFound {
-		return &resource, nil
-	}
-	return nil, fmt.Errorf("invalid resource type %s", resourceType)
-}
-
-// Retrieve the resource slug / shortname based on the resourceType and the selected convention
-func getSlug(resourceType string, convention string) string {
-	if convention == ConventionCafClassic || convention == ConventionCafRandom {
-		if val, ok := ResourceDefinitions[resourceType]; ok {
-			return val.CafPrefix
-		}
-	}
-	return ""
-}
-
-func trimResourceName(resourceName string, maxLength int) string {
-	var length int = len(resourceName)
-
-	if length > maxLength {
-		length = maxLength
-	}
-
-	return string(resourceName[0:length])
-}
-
-func convertInterfaceToString(source []interface{}) []string {
-	s := make([]string, len(source))
-	for i, v := range source {
-		s[i] = fmt.Sprint(v)
-	}
-	return s
-}
-
-func composeName(separator string,
-	prefixes []string,
-	name string,
-	slug string,
-	suffixes []string,
-	randomSuffix string,
-	maxlength int,
-	namePrecedence []string) string {
-	contents := []string{}
-	currentlength := 0
-
-	for i := 0; i < len(namePrecedence); i++ {
-		initialized := 0
-		if len(contents) > 0 {
-			initialized = len(separator)
-		}
-		switch c := namePrecedence[i]; c {
-		case "name":
-			if len(name) > 0 {
-				if currentlength+len(name)+initialized <= maxlength {
-					contents = append(contents, name)
-					currentlength = currentlength + len(name) + initialized
-				}
-			}
-		case "slug":
-			if len(slug) > 0 {
-				if currentlength+len(slug)+initialized <= maxlength {
-					contents = append([]string{slug}, contents...)
-					currentlength = currentlength + len(slug) + initialized
-				}
-			}
-		case "random":
-			if len(randomSuffix) > 0 {
-				if currentlength+len(randomSuffix)+initialized <= maxlength {
-					contents = append(contents, randomSuffix)
-					currentlength = currentlength + len(randomSuffix) + initialized
-				}
-			}
-		case "suffixes":
-			if len(suffixes) > 0 {
-				if len(suffixes[0]) > 0 {
-					if currentlength+len(suffixes[0])+initialized <= maxlength {
-						contents = append(contents, suffixes[0])
-						currentlength = currentlength + len(suffixes[0]) + initialized
-					}
-				}
-				suffixes = suffixes[1:]
-				if len(suffixes) > 0 {
-					i--
-				}
-			}
-		case "prefixes":
-			if len(prefixes) > 0 {
-				if len(prefixes[len(prefixes)-1]) > 0 {
-					if currentlength+len(prefixes[len(prefixes)-1])+initialized <= maxlength {
-						contents = append([]string{prefixes[len(prefixes)-1]}, contents...)
-						currentlength = currentlength + len(prefixes[len(prefixes)-1]) + initialized
-					}
-				}
-				prefixes = prefixes[:len(prefixes)-1]
-				if len(prefixes) > 0 {
-					i--
-				}
-			}
-
-		}
-
-	}
-	content := strings.Join(contents, separator)
-	return content
-}
-
-func validateResourceType(resourceType string, resourceTypes []string) (bool, error) {
-	isEmpty := len(resourceType) == 0 && len(resourceTypes) == 0
-	if isEmpty {
-		return false, fmt.Errorf("resource_type and resource_types parameters are empty, you must specify at least one resource type")
-	}
-	errorStrings := []string{}
-	resourceList := resourceTypes
-	if len(resourceType) > 0 {
-		resourceList = append(resourceList, resourceType)
-	}
-
-	for _, resource := range resourceList {
-		_, err := getResource(resource)
-		if err != nil {
-			errorStrings = append(errorStrings, err.Error())
-		}
-	}
-	if len(errorStrings) > 0 {
-		return false, fmt.Errorf(strings.Join(errorStrings, "\n"))
-	}
-	return true, nil
-}
-
-func getResourceName(resourceTypeName string, separator string,
-	prefixes []string,
-	name string,
-	suffixes []string,
-	randomSuffix string,
-	convention string,
-	cleanInput bool,
-	passthrough bool,
-	useSlug bool,
-	namePrecedence []string) (string, error) {
-
-	resource, err := getResource(resourceTypeName)
-	if err != nil {
-		return "", err
-	}
-	validationRegEx, err := regexp.Compile(resource.ValidationRegExp)
-	if err != nil {
-		return "", err
-	}
-
-	slug := ""
-	if useSlug {
-		slug = getSlug(resourceTypeName, convention)
-	}
-
-	if cleanInput {
-		prefixes = cleanSlice(prefixes, resource)
-		suffixes = cleanSlice(suffixes, resource)
-		name = cleanString(name, resource)
-		separator = cleanString(separator, resource)
-		randomSuffix = cleanString(randomSuffix, resource)
-	}
-
-	var resourceName string
-
-	if passthrough {
-		resourceName = name
-	} else {
-		resourceName = composeName(separator, prefixes, name, slug, suffixes, randomSuffix, resource.MaxLength, namePrecedence)
-	}
-	resourceName = trimResourceName(resourceName, resource.MaxLength)
-
-	if resource.LowerCase {
-		resourceName = strings.ToLower(resourceName)
-	}
-
-	if !validationRegEx.MatchString(resourceName) {
-		return "", fmt.Errorf("invalid name for CAF naming %s %s, the pattern %s doesn't match %s", resource.ResourceTypeName, name, resource.ValidationRegExp, resourceName)
-	}
-
-	return resourceName, nil
-}
-
-func getNameResult(d *schema.ResourceData, meta interface{}) error {
+func getDifference(context context.Context, d *schema.ResourceDiff, resource interface{}) error {
 	name := d.Get("name").(string)
 	prefixes := convertInterfaceToString(d.Get("prefixes").([]interface{}))
 	suffixes := convertInterfaceToString(d.Get("suffixes").([]interface{}))
@@ -456,28 +156,93 @@ func getNameResult(d *schema.ResourceData, meta interface{}) error {
 
 	randomSuffix := randSeq(int(randomLength), &randomSeed)
 	namePrecedence := []string{"name", "slug", "random", "suffixes", "prefixes"}
-
-	isValid, err := validateResourceType(resourceType, resourceTypes)
-	if !isValid {
+	result, results, id, err :=
+		getData(resourceType, resourceTypes, separator, prefixes, name, suffixes, randomSuffix, convention, cleanInput, passthrough, useSlug, namePrecedence)
+	if err != nil {
 		return err
 	}
-
-	if len(resourceType) > 0 {
-		resourceName, err := getResourceName(resourceType, separator, prefixes, name, suffixes, randomSuffix, convention, cleanInput, passthrough, useSlug, namePrecedence)
-		if err != nil {
-			return err
-		}
-		d.Set("result", resourceName)
-	}
-	resourceNames := make(map[string]string, len(resourceTypes))
-	for _, resourceTypeName := range resourceTypes {
-		var err error
-		resourceNames[resourceTypeName], err = getResourceName(resourceTypeName, separator, prefixes, name, suffixes, randomSuffix, convention, cleanInput, passthrough, useSlug, namePrecedence)
-		if err != nil {
-			return err
-		}
-	}
-	d.Set("results", resourceNames)
-	d.SetId(randSeq(16, nil))
+	log.Debugf("%v,%v,%v", result, results, id)
 	return nil
+}
+
+func resourceNameCreate(d *schema.ResourceData, meta interface{}) error {
+	return getNameResult(d, meta)
+}
+
+func resourceNameUpdate(d *schema.ResourceData, meta interface{}) error {
+	return getNameResult(d, meta)
+}
+
+func resourceNameRead(d *schema.ResourceData, meta interface{}) error {
+	return nil
+}
+
+func convertInterfaceToString(source []interface{}) []string {
+	s := make([]string, len(source))
+	for i, v := range source {
+		s[i] = fmt.Sprint(v)
+	}
+	return s
+}
+
+func getNameResult(d *schema.ResourceData, meta interface{}) error {
+	log.Infof("Get Name result, %v", d)
+	name := d.Get("name").(string)
+	prefixes := convertInterfaceToString(d.Get("prefixes").([]interface{}))
+	suffixes := convertInterfaceToString(d.Get("suffixes").([]interface{}))
+	separator := d.Get("separator").(string)
+	resourceType := d.Get("resource_type").(string)
+	resourceTypes := convertInterfaceToString(d.Get("resource_types").([]interface{}))
+	cleanInput := d.Get("clean_input").(bool)
+	passthrough := d.Get("passthrough").(bool)
+	useSlug := d.Get("use_slug").(bool)
+	randomLength := d.Get("random_length").(int)
+	randomSeed := int64(d.Get("random_seed").(int))
+
+	convention := ConventionCafClassic
+
+	randomSuffix := randSeq(int(randomLength), &randomSeed)
+	namePrecedence := []string{"name", "slug", "random", "suffixes", "prefixes"}
+	result, results, id, err :=
+		getData(resourceType, resourceTypes, separator, prefixes, name, suffixes, randomSuffix, convention, cleanInput, passthrough, useSlug, namePrecedence)
+	if err != nil {
+		return err
+	}
+	if len(result) > 0 {
+		d.Set("result", result)
+	}
+	if len(results) > 0 {
+		d.Set("results", results)
+	}
+	d.SetId(id)
+	return nil
+}
+
+func getData(resourceType string, resourceTypes []string, separator string, prefixes []string, name string, suffixes []string, randomSuffix string, convention string, cleanInput bool, passthrough bool, useSlug bool, namePrecedence []string) (result string, results map[string]string, id string, err error) {
+	isValid, err := validateResourceType(resourceType, resourceTypes)
+	if !isValid {
+		return
+	}
+	if results == nil {
+		results = make(map[string]string)
+	}
+	ids := []string{}
+	if len(resourceType) > 0 {
+		result, err = getResourceName(resourceType, separator, prefixes, name, suffixes, randomSuffix, convention, cleanInput, passthrough, useSlug, namePrecedence)
+		if err != nil {
+			return
+		}
+		results[resourceType] = result
+		ids = append(ids, fmt.Sprintf("%s\t%s", resourceType, result))
+	}
+
+	for _, resourceTypeName := range resourceTypes {
+		results[resourceTypeName], err = getResourceName(resourceTypeName, separator, prefixes, name, suffixes, randomSuffix, convention, cleanInput, passthrough, useSlug, namePrecedence)
+		if err != nil {
+			return
+		}
+		ids = append(ids, fmt.Sprintf("%s\t%s", resourceTypeName, results[resourceTypeName]))
+	}
+	id = b64.StdEncoding.EncodeToString([]byte(strings.Join(ids, "\n")))
+	return
 }
