+++ azurecaf/resource_name_engine.go
@@ -1,0 +1,209 @@
+package azurecaf
+
+import (
+	"fmt"
+	"regexp"
+	"strings"
+)
+
+func cleanSlice(names []string, resourceDefinition *ResourceStructure) []string {
+	for i, name := range names {
+		names[i] = cleanString(name, resourceDefinition)
+	}
+	return names
+}
+
+func cleanString(name string, resourceDefinition *ResourceStructure) string {
+	myRegex, _ := regexp.Compile(resourceDefinition.RegEx)
+	return myRegex.ReplaceAllString(name, "")
+}
+
+func concatenateParameters(separator string, parameters ...[]string) string {
+	elems := []string{}
+	for _, items := range parameters {
+		for _, item := range items {
+			if len(item) > 0 {
+				elems = append(elems, []string{item}...)
+			}
+		}
+	}
+	return strings.Join(elems, separator)
+}
+
+func getResource(resourceType string) (*ResourceStructure, error) {
+	if resourceKey, existing := ResourceMaps[resourceType]; existing {
+		resourceType = resourceKey
+	}
+	if resource, resourceFound := ResourceDefinitions[resourceType]; resourceFound {
+		return &resource, nil
+	}
+	return nil, fmt.Errorf("invalid resource type %s", resourceType)
+}
+
+// Retrieve the resource slug / shortname based on the resourceType and the selected convention
+func getSlug(resourceType string, convention string) string {
+	if convention == ConventionCafClassic || convention == ConventionCafRandom {
+		if val, ok := ResourceDefinitions[resourceType]; ok {
+			return val.CafPrefix
+		}
+	}
+	return ""
+}
+
+func trimResourceName(resourceName string, maxLength int) string {
+	var length int = len(resourceName)
+
+	if length > maxLength {
+		length = maxLength
+	}
+
+	return string(resourceName[0:length])
+}
+
+func composeName(separator string,
+	prefixes []string,
+	name string,
+	slug string,
+	suffixes []string,
+	randomSuffix string,
+	maxlength int,
+	namePrecedence []string) string {
+	contents := []string{}
+	currentlength := 0
+
+	for i := 0; i < len(namePrecedence); i++ {
+		initialized := 0
+		if len(contents) > 0 {
+			initialized = len(separator)
+		}
+		switch c := namePrecedence[i]; c {
+		case "name":
+			if len(name) > 0 {
+				if currentlength+len(name)+initialized <= maxlength {
+					contents = append(contents, name)
+					currentlength = currentlength + len(name) + initialized
+				}
+			}
+		case "slug":
+			if len(slug) > 0 {
+				if currentlength+len(slug)+initialized <= maxlength {
+					contents = append([]string{slug}, contents...)
+					currentlength = currentlength + len(slug) + initialized
+				}
+			}
+		case "random":
+			if len(randomSuffix) > 0 {
+				if currentlength+len(randomSuffix)+initialized <= maxlength {
+					contents = append(contents, randomSuffix)
+					currentlength = currentlength + len(randomSuffix) + initialized
+				}
+			}
+		case "suffixes":
+			if len(suffixes) > 0 {
+				if len(suffixes[0]) > 0 {
+					if currentlength+len(suffixes[0])+initialized <= maxlength {
+						contents = append(contents, suffixes[0])
+						currentlength = currentlength + len(suffixes[0]) + initialized
+					}
+				}
+				suffixes = suffixes[1:]
+				if len(suffixes) > 0 {
+					i--
+				}
+			}
+		case "prefixes":
+			if len(prefixes) > 0 {
+				if len(prefixes[len(prefixes)-1]) > 0 {
+					if currentlength+len(prefixes[len(prefixes)-1])+initialized <= maxlength {
+						contents = append([]string{prefixes[len(prefixes)-1]}, contents...)
+						currentlength = currentlength + len(prefixes[len(prefixes)-1]) + initialized
+					}
+				}
+				prefixes = prefixes[:len(prefixes)-1]
+				if len(prefixes) > 0 {
+					i--
+				}
+			}
+
+		}
+
+	}
+	content := strings.Join(contents, separator)
+	return content
+}
+
+func validateResourceType(resourceType string, resourceTypes []string) (bool, error) {
+	isEmpty := len(resourceType) == 0 && len(resourceTypes) == 0
+	if isEmpty {
+		return false, fmt.Errorf("resource_type and resource_types parameters are empty, you must specify at least one resource type")
+	}
+	errorStrings := []string{}
+	resourceList := resourceTypes
+	if len(resourceType) > 0 {
+		resourceList = append(resourceList, resourceType)
+	}
+
+	for _, resource := range resourceList {
+		_, err := getResource(resource)
+		if err != nil {
+			errorStrings = append(errorStrings, err.Error())
+		}
+	}
+	if len(errorStrings) > 0 {
+		return false, fmt.Errorf(strings.Join(errorStrings, "\n"))
+	}
+	return true, nil
+}
+
+func getResourceName(resourceTypeName string, separator string,
+	prefixes []string,
+	name string,
+	suffixes []string,
+	randomSuffix string,
+	convention string,
+	cleanInput bool,
+	passthrough bool,
+	useSlug bool,
+	namePrecedence []string) (string, error) {
+
+	resource, err := getResource(resourceTypeName)
+	if err != nil {
+		return "", err
+	}
+	validationRegEx, err := regexp.Compile(resource.ValidationRegExp)
+	if err != nil {
+		return "", err
+	}
+
+	slug := ""
+	if useSlug {
+		slug = getSlug(resourceTypeName, convention)
+	}
+
+	if cleanInput {
+		prefixes = cleanSlice(prefixes, resource)
+		suffixes = cleanSlice(suffixes, resource)
+		name = cleanString(name, resource)
+		separator = cleanString(separator, resource)
+		randomSuffix = cleanString(randomSuffix, resource)
+	}
+
+	var resourceName string
+
+	if passthrough {
+		resourceName = name
+	} else {
+		resourceName = composeName(separator, prefixes, name, slug, suffixes, randomSuffix, resource.MaxLength, namePrecedence)
+	}
+	resourceName = trimResourceName(resourceName, resource.MaxLength)
+
+	if resource.LowerCase {
+		resourceName = strings.ToLower(resourceName)
+	}
+
+	if !validationRegEx.MatchString(resourceName) {
+		return "", fmt.Errorf("invalid name for CAF naming %s %s, the pattern %s doesn't match %s", resource.ResourceTypeName, name, resource.ValidationRegExp, resourceName)
+	}
+
+	return resourceName, nil
+}
