--- azurecaf/resource_naming_convention_test.go
@@ -1,56 +1,0 @@
-package azurecaf
-
-import (
-	"fmt"
-	"regexp"
-	"strings"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
-)
-
-func testAccCafNamingValidation(id string, name string, expectedLength int, prefix string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		rs, ok := s.RootModule().Resources[id]
-		if !ok {
-			return fmt.Errorf("Not found: %s", id)
-		}
-		if rs.Primary.ID == "" {
-			return fmt.Errorf("No ID is set")
-		}
-
-		attrs := rs.Primary.Attributes
-
-		result := attrs["result"]
-		if len(result) != expectedLength {
-			return fmt.Errorf("got %s %d result items; want %d", result, len(result), expectedLength)
-		}
-		if !strings.HasPrefix(result, prefix) {
-			return fmt.Errorf("got %s which doesn't start with %s", result, prefix)
-		}
-		if !strings.Contains(result, name) {
-			return fmt.Errorf("got %s which doesn't contain the name %s", result, name)
-		}
-		return nil
-	}
-}
-
-func regexMatch(id string, exp *regexp.Regexp, requiredMatches int) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		rs, ok := s.RootModule().Resources[id]
-		if !ok {
-			return fmt.Errorf("Not found: %s", id)
-		}
-		if rs.Primary.ID == "" {
-			return fmt.Errorf("No ID is set")
-		}
-
-		result := rs.Primary.Attributes["result"]
-
-		if matches := exp.FindAllStringSubmatchIndex(result, -1); len(matches) != requiredMatches {
-			return fmt.Errorf("result string is %s; did not match %s, got %d", result, exp, len(matches))
-		}
-
-		return nil
-	}
-}
