+++ go.mod
@@ -2,4 +2,7 @@ module github.com/aztfmod/terraform-provider-azurecaf
 
 go 1.15
 
-require github.com/hashicorp/terraform-plugin-sdk/v2 v2.6.1
+require (
+	github.com/google/martian/v3 v3.1.0
+	github.com/hashicorp/terraform-plugin-sdk/v2 v2.6.1
+)
